refactor(storage): extract PurgeOperationStoreMixin from governance (SQLite + ABC)
Move the PurgeOperationStore bucket out of the residual GovernanceMixin into a
new PurgeOperationStoreMixin in the governance/ package, across the SQLite
backend and the backend-neutral ABC. Behaviour-preserving, verbatim move.

The bucket: begin_purge_operation, record_purge_target, list_purge_targets,
purge_targets_prepared, prepare_governance_erase_targets, fail_purge_operation,
get_purge_operation, plus the Purge-owned private _record_purge_target_locked
(consumed cross-bucket by the residual rebuild-hide / erase-execution / barrier
methods, which reach it via MRO co-composition).

Both mixins stay co-composed on the concrete Storage class. The residual keeps
pure-annotation stubs for the cross-bucket privates it calls
(get_purge_operation, _record_purge_target_locked); the new mixin keeps stubs
for the shared infra and the cross-bucket residents it calls
(_owned_user_playbook_ids_locked, _planned_governance_delete_counts). Shared
row/JSON helpers stay module-level in the residual and are imported.

Repoint the one test that patches _epoch_now for fail_purge_operation from the
residual _governance module to the governance._purge module.
