+++ reflexio/server/services/storage/storage_base/governance/__init__.py
@@ -1,5 +1,7 @@
 from ._audit import AuditEventStoreMixin
+from ._purge import PurgeOperationStoreMixin
 
 __all__ = [
     "AuditEventStoreMixin",
+    "PurgeOperationStoreMixin",
 ]
